requirements.txt
openai
python-dotenv
pyyaml
requests